requirements.txt
argh==0.26.2
asgiref==3.4.1
baton==1.0.1
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
curlify==2.2.1
DateTime==4.3
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-baton==2.2.3
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
docker-py==1.9.0
future==0.18.2
hgicommon==1.2.0
hgijson==1.3.1
idna==3.2
inflection==0.5.1
jmespath==0.10.0
mysqlclient==2.0.3
oauthlib==3.1.1
pathtools==0.1.2
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.5.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
watchdog==0.8.3
websocket-client==1.2.1
zope.interface==5.4.0
